Vulkan: Minor warnings fixes in Vulkan code

+++ Hermes/Source/RenderInterface/Vulkan/VulkanRenderPass.cpp
@@ -218,7 +218,7 @@ namespace Hermes
 				}
 
 				NewSubpass.pPreserveAttachments = Subpass.PreservedAttachmentsIndices.data();
-				NewSubpass.preserveAttachmentCount = Subpass.PreservedAttachmentsIndices.size();
+				NewSubpass.preserveAttachmentCount = (uint32)Subpass.PreservedAttachmentsIndices.size();
 				
 				VulkanSubpasses.push_back(NewSubpass);
 			}
